feat: adds AppendRequestBody.
Currently we were blindly copying the request body from the upstream call to the
RequestBodyWriter and then back to the downstream handler in many cases wasting resources
because body was beyong what coraza should analyse and or request end up being rejected
due to the size. This method allows to early check conditions around the body size to be
able to resolve a rejecting before copying the the body (or part of it).