breaking: drops content injection features.
Currently we support SecContentInjection and the actions around it, however there is a general feeling that manipulating the body is not something WAF as core should do, instead connectors should take care of that. This is specially inconvenient depending on the connectors lifecycle. For example, in proxy-wasm we can't modify the size of the response body as by the time we want to modify or resolve a phase 4, the headers have been sent already and hence the content length. It is better to drop this feature (only supported in Modsec 2.5) rather han having a half baked solution.